# openapi.yaml
openapi: 3.0.3
info:
  title: cerbix API
  version: 0.19.0
  description: >
    Internal uptime & SLA monitoring API. Authentication is via a server-side
    session cookie established by an OIDC provider (`/auth/login` → `/auth/callback`)
    or local login (`/auth/local/login`). All `/api/v1/*` endpoints require a
    valid session and enforce tenant isolation: callers only see and mutate the
    organizations/projects they are authorized for (non-members receive 404;
    insufficient in-org role receives 403).
servers:
  - url: /
    description: Same-origin (SPA is served by the API binary)
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/v1/projects/{projectID}/monitors:
    parameters: [ { $ref: '#/components/parameters/ProjectID' } ]
    get:
      tags: [monitors]
      summary: List a project's monitors
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Monitor' } }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [monitors]
      summary: Create a monitor (editor+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMonitor' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Monitor' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/monitors/{monitorID}:
    parameters: [ { $ref: '#/components/parameters/MonitorID' } ]
    get:
      tags: [monitors]
      summary: Get a monitor
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Monitor' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [monitors]
      summary: Update a monitor's mutable fields (editor+). Type is immutable.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMonitor' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Monitor' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [monitors]
      summary: Delete a monitor (editor+)
      responses:
        '204': { description: Deleted. }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cerbix_session
    bearerAuth:
      type: http
      scheme: bearer
      description: "Machine auth via Authorization: Bearer. Either a cerbix service-account token (cbx_...) or an OIDC client-credentials access token (a realm-issued JWT, when OIDC is enabled); the JWT subject is JIT-provisioned and access is membership-gated."
  parameters:
...
    ProjectID:
      in: path
      name: projectID
      required: true
      schema: { type: string, format: uuid }
...
    MonitorID:
      in: path
      name: monitorID
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: Authenticated but insufficient role.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found or not visible to the caller (existence hidden).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Monitor:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        name: { type: string }
        type: { type: string, enum: [http, tcp, icmp, dns, tls, grpc, composite, postgres, mysql, redis, promql, rabbitmq, websocket, ssh, synthetic, push] }
        target: { type: string }
        method: { type: string, enum: [GET, POST, HEAD, PUT, DELETE], description: 'HTTP request method (http monitors).' }
        interval_seconds: { type: integer }
        timeout_seconds: { type: integer }
        retries: { type: integer }
        failure_threshold: { type: integer, description: 'Consecutive failed checks before "down" (confirmations); default 1.' }
        confirm_interval_seconds: { type: integer, description: 'Accelerated probe interval while confirming a failure (first fail → verdict). 0 = off; clamped to [5, interval]; effective with failure_threshold > 1.' }
        depends_on:
          type: array
          items: { type: string, format: uuid }
          description: 'Parent monitors (same project). While any transitive parent is down, alerts for this monitor are suppressed; data keeps recording. Must stay acyclic.'
        consecutive_failures: { type: integer, readOnly: true, description: 'Live confirmation counter ("confirming N/M" when > 0 and status is up).' }
        renotify_seconds: { type: integer, description: 'Re-send the down alert every N seconds while down (0 = off).' }
        grace_seconds: { type: integer, description: 'Push monitors: extra tolerance beyond the interval before "down".' }
        conditions: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string }, description: 'Free-form labels (env:prod, team:x) for filtering.' }
        region: { type: string, description: 'Worker-pool region that probes this monitor; default core. Composite monitors must be core.' }
        config: { type: object, additionalProperties: { type: string }, description: 'Per-type config; for composite: {children:"id1,id2", mode:"all|any"}.' }
        enabled: { type: boolean }
        auto_incident: { type: boolean, description: 'When true, an incident opens automatically when this monitor goes down.' }
        escalation_policy_id: { type: string, description: 'On-call escalation policy for down alerts. Empty = flat notification to all linked channels.' }
        status: { type: string, enum: [pending, up, down] }
        push_token: { type: string, description: 'Set for push monitors; the secret in their heartbeat URL.' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateMonitor:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        type: { type: string, enum: [http, tcp, icmp, dns, tls, grpc, composite, postgres, mysql, redis, promql, rabbitmq, websocket, ssh, synthetic, push] }
        target: { type: string }
        method: { type: string, enum: [GET, POST, HEAD, PUT, DELETE], description: 'HTTP request method; defaults to GET.' }
        interval_seconds: { type: integer, default: 60 }
        timeout_seconds: { type: integer, default: 10 }
        retries: { type: integer, default: 0 }
        failure_threshold: { type: integer, default: 1, description: 'Consecutive failed checks before "down".' }
        confirm_interval_seconds: { type: integer, default: 10, description: 'Faster probes while confirming a failure. 0 = off; clamped to [5, interval].' }
        depends_on:
          type: array
          items: { type: string, format: uuid }
          description: 'Parent monitors of the same project; cycles are rejected (400).'
        renotify_seconds: { type: integer, default: 0, description: 'Re-send the down alert every N seconds while down (0 = off).' }
        grace_seconds: { type: integer, default: 0, description: 'Push monitors: extra tolerance beyond the interval.' }
        conditions:
          type: array
          items: { type: string }
          example: ["[STATUS] == 200", "[RESPONSE_TIME] < 500"]
        tags: { type: array, items: { type: string }, description: 'Free-form labels (env:prod, team:x) for filtering.' }
        region: { type: string, description: 'Worker-pool region that probes this monitor; default core. Composite monitors must be core.' }
        config: { type: object, additionalProperties: { type: string }, description: 'Per-type config; for composite: {children:"id1,id2", mode:"all|any"}.' }
        enabled: { type: boolean, default: true }
        auto_incident: { type: boolean, default: true, description: 'Open an incident automatically when this monitor goes down.' }
        escalation_policy_id: { type: string, description: 'On-call escalation policy for down alerts (requires auto_incident). Empty = flat notification.' }
    UpdateMonitor:
      type: object
      description: Partial update; omitted fields are unchanged. Type is immutable.
      properties:
        name: { type: string }
        target: { type: string }
        method: { type: string, enum: [GET, POST, HEAD, PUT, DELETE] }
        interval_seconds: { type: integer }
        timeout_seconds: { type: integer }
        retries: { type: integer }
        failure_threshold: { type: integer }
        confirm_interval_seconds: { type: integer }
        depends_on:
          type: array
          items: { type: string, format: uuid }
        renotify_seconds: { type: integer }
        grace_seconds: { type: integer }
        conditions: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string }, description: 'Free-form labels (env:prod, team:x) for filtering.' }
        region: { type: string, description: 'Worker-pool region that probes this monitor; default core. Composite monitors must be core.' }
        config: { type: object, additionalProperties: { type: string }, description: 'Per-type config; for composite: {children:"id1,id2", mode:"all|any"}.' }
        enabled: { type: boolean }
        auto_incident: { type: boolean }
        escalation_policy_id: { type: string }
...